+++ ligo/skymap/plot/marker.py
@@ -19,8 +19,9 @@
 import numpy as np
 from matplotlib.markers import MarkerStyle
 from matplotlib.path import Path
+from matplotlib.transforms import Affine2D
 
-__all__ = ('earth', 'sun', 'reticle')
+__all__ = ('earth', 'sun', 'moon', 'reticle')
 
 
 earth = Path.unit_circle()
@@ -75,6 +76,79 @@ Examples
     plt.plot(0, 0, marker=sun, markersize=20, markeredgewidth=2)
 
 """
+
+
+def moon(phase, shadow=False):
+    """Create a marker in the shape of the Moon.
+
+    Parameters
+    ----------
+    phase : float
+        Lunar phase in degrees between -180 and 180.
+    shadow : bool
+        If set, then the shadowed portion of the Moon is included in the
+        marker, and its fill color can be set independently using the
+        ``markerfacecoloralt`` keyword argument for
+        :meth:`~matplotlib.axes.Axes.plot` (see
+        :ref:`matplotlib:marker_fill_styles`).
+
+    Returns
+    -------
+    markerstyle : matplotlib.markers.MarkerStyle
+
+    Examples
+    --------
+
+    .. plot::
+       :context: reset
+       :include-source:
+       :align: center
+
+        from matplotlib import pyplot as plt
+        from matplotlib.ticker import MultipleLocator
+        import numpy as np
+        from ligo.skymap.plot.marker import moon
+
+        d_phase = 30
+        phases = np.arange(-180, 180 + d_phase, d_phase)
+
+        fig, ax = plt.subplots(figsize=(8, 3), tight_layout=True)
+        ax.xaxis.set_major_locator(MultipleLocator(d_phase))
+        for phase in phases:
+            ax.plot(phase, 4, ms=20, marker=moon(phase, shadow=False), mfc="none", mec="black")
+            ax.plot(phase, 3, ms=20, marker=moon(phase, shadow=False), mfc="goldenrod", mec="none")
+            ax.plot(phase, 2, ms=20, marker=moon(phase, shadow=False), mfc="goldenrod", mec="k")
+            ax.plot(phase, 1, ms=20, marker=moon(phase, shadow=True), mfc="goldenrod", mfcalt="gray", mec="none")
+            ax.plot(phase, 0, ms=20, marker=moon(phase, shadow=True), mfc="goldenrod", mfcalt="gray", mec="black")
+        ax.set_yticks(
+            [0, 1, 2, 3, 4],
+            ["shadow, fill, stroke", "shadow, fill", "fill, stroke", "fill", "stroke"],
+        )
+        ax.set_ylim(-0.5, 4.5)
+
+    """  # noqa: E501
+    angle = np.deg2rad(90 - phase)
+    sign = np.sign(np.cos(angle))
+    arc = Path.arc(90, 270, 9)
+
+    path1 = arc.transformed(Affine2D().scale(sign * np.sin(angle), 1))
+    path2 = arc.transformed(Affine2D().scale(-sign, 1))
+    path3 = arc.transformed(Affine2D().scale(sign, 1))
+
+    light_path = Path(
+        np.concatenate((path1.vertices, path2.vertices[::-1])),
+        np.concatenate((path1.codes, path2.codes[:0:-1], [Path.CLOSEPOLY])),
+    )
+    dark_path = Path(
+        np.concatenate((path1.vertices, path3.vertices[::-1])),
+        np.concatenate((path1.codes, path3.codes[:0:-1], [Path.CLOSEPOLY])),
+    )
+
+    markerstyle = MarkerStyle(light_path, joinstyle='miter')
+    if shadow:
+        markerstyle._alt_path = dark_path
+        markerstyle._alt_transform = markerstyle._transform
+    return markerstyle
 
 
 def reticle(inner=0.5, outer=1.0, angle=0.0, which='lrtb'):